Refactor MidiReader tests to be more explicit
Previously, all MidiReader tests used an external fixture file. Because
the contents of this file are not visible among the tests (and because
it is not human-readable, besides), it was difficult to understand if
the tests were correct. This commit updates tests to use explicit binary
strings wherever possible, making it easier to verify test correctness.